test: fix the testFuzz_WhenAmountToClaimIsGreaterThenZero_Failing #748 (#758)
* fix: the testFuzz_WhenAmountToClaimIsGreaterThenZero_Failing #748

* refactor: unecessary spaces #748

* refactor: rename testFuzz_WhenAmountToClaimIsGreaterThenZero

+++ src/market-making/branches/VaultRouterBranch.sol
@@ -649,7 +649,7 @@ contract VaultRouterBranch {
     )
         external
         view
-        returns (uint128 totalShares, int128 valuePerShare, uint128 accountShares, int128 accountLastValuePerShare)
+        returns (uint128 totalShares, int256 valuePerShare, uint128 accountShares, int256 accountLastValuePerShare)
     {
         // get vault staking fee distribution data
         Distribution.Data storage distributionData = Vault.loadLive(vaultId).wethRewardDistribution;
